pyproject.toml
[tool.poetry]
name = "deepseek"
version = "0.1.0"
description = ""
authors = ["Dennis Seah <dennis.seah@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
lagom = "^2.7.5"
python-dotenv = "^1.0.1"
azure-core = "^1.32.0"
azure-identity = "^1.19.0"
azure-ai-inference = "^1.0.0b8"
pydantic = "^2.10.6"
aiohttp = "^3.11.12"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"
pyright = "^1.1.393"
pre-commit = "^4.1.0"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "az_deepseek/hosting.py",
    "az_deepseek/protocols/*"
]